requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
Markdown==3.5.1
psycopg==3.1.13
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.3
